Update: moved from scp to rsync.
Seems like OpenSSH is considering deprecating spc and they suggest sftp and rsync as replacements (sftp is part of the OpenSSH package). People say rsync is better so I'm using rsync. It is also an improvement to deploy times/sizes because it skips files that are the same on the remote.