refactor: Rename Blog.date to createDate and add updateDate
Renames the `date` field in the `Blog` model to `createDate` to
better reflect its purpose. Also, adds an `updateDate` field to
track when a blog post was last updated.

This change improves clarity and provides more accurate
information about the blog posts.

Updates the UI to use the new `createDate` field for displaying
the blog post date.

Signed-off-by: JoaoCostaIFG <joaocosta.work@posteo.net>